+++ src/app/pages/dashboard/components/linea-tiempo/timelinewidget.html
@@ -7,19 +7,4 @@
             {{ module.moment}}
         </ng-template>
     </p-timeline>
-
-    <!--<p-timeline [value]="events" layout="horizontal" align="bottom">
-        <ng-template #content let-event>
-            {{ event }}
-        </ng-template>
-
-    </p-timeline>
-    <p-timeline [value]="events" layout="horizontal" align="alternate">
-        <ng-template #content let-event>
-            {{ event }}
-        </ng-template>
-        <ng-template #opposite let-event>
-            <span>&nbsp;</span>
-        </ng-template>
-    </p-timeline>-->
 </div>
