+++ src/app/pages/dashboard/components/notificaciones/notificationswidget.html
@@ -1,0 +1,68 @@
+<div class="card">
+    <div class="flex items-center justify-between mb-6">
+        <div class="font-semibold text-xl">Notificaciones</div>
+    </div>
+
+    <span class="block text-muted-color font-medium mb-4">HOY</span>
+    <ul class="p-0 mx-0 mt-0 mb-6 list-none">
+        <li class="flex items-center py-2 border-b border-surface">
+            <div
+                class="w-12 h-12 flex items-center justify-center bg-blue-100 dark:bg-blue-400/10 rounded-full mr-4 shrink-0">
+                <i class="pi pi-question text-xl! text-blue-500"></i>
+            </div>
+            <span class="text-surface-900 dark:text-surface-0 leading-normal">Tutor asignado
+                <span class="text-surface-700 dark:text-surface-100">Notificación 1</span>
+            </span>
+        </li>
+        <li class="flex items-center py-2">
+            <div
+                class="w-12 h-12 flex items-center justify-center bg-orange-100 dark:bg-orange-400/10 rounded-full mr-4 shrink-0">
+                <i class="pi pi-download text-xl! text-orange-500"></i>
+            </div>
+            <span class="text-surface-700 dark:text-surface-100 leading-normal">Tutor asignado <span
+                    class="text-primary font-bold"></span> Notificación 2</span>
+        </li>
+    </ul>
+
+    <span class="block text-muted-color font-medium mb-4">AYER</span>
+    <ul class="p-0 m-0 list-none mb-6">
+        <li class="flex items-center py-2 border-b border-surface">
+            <div
+                class="w-12 h-12 flex items-center justify-center bg-blue-100 dark:bg-blue-400/10 rounded-full mr-4 shrink-0">
+                <i class="pi pi-question text-xl! text-blue-500"></i>
+            </div>
+            <span class="text-surface-900 dark:text-surface-0 leading-normal">Tutor asignado
+                <span class="text-surface-700 dark:text-surface-100">Notificación 3 <span
+                        class="text-primary font-bold"></span></span>
+            </span>
+        </li>
+        <li class="flex items-center py-2 border-b border-surface">
+            <div
+                class="w-12 h-12 flex items-center justify-center bg-pink-100 dark:bg-pink-400/10 rounded-full mr-4 shrink-0">
+                <i class="pi pi-question text-xl! text-pink-500"></i>
+            </div>
+            <span class="text-surface-900 dark:text-surface-0 leading-normal">Tutor asignado
+                <span class="text-surface-700 dark:text-surface-100">Notificación 4</span>
+            </span>
+        </li>
+    </ul>
+    <span class="block text-muted-color font-medium mb-4">SEMANA PASADA</span>
+    <ul class="p-0 m-0 list-none">
+        <li class="flex items-center py-2 border-b border-surface">
+            <div
+                class="w-12 h-12 flex items-center justify-center bg-green-100 dark:bg-green-400/10 rounded-full mr-4 shrink-0">
+                <i class="pi pi-arrow-up text-xl! text-green-500"></i>
+            </div>
+            <span class="text-surface-900 dark:text-surface-0 leading-normal">Tutor asignado <span
+                    class="text-primary font-bold"></span>.</span>
+        </li>
+        <li class="flex items-center py-2 border-b border-surface">
+            <div
+                class="w-12 h-12 flex items-center justify-center bg-purple-100 dark:bg-purple-400/10 rounded-full mr-4 shrink-0">
+                <i class="pi pi-heart text-xl! text-purple-500"></i>
+            </div>
+            <span class="text-surface-900 dark:text-surface-0 leading-normal"><span
+                    class="text-primary font-bold"></span> Tutor asignado</span>
+        </li>
+    </ul>
+</div>
