+++ src/app/pages/dashboard/components/notificaciones/notificationswidget.ts
@@ -1,0 +1,13 @@
+import { Component } from '@angular/core';
+import { ButtonModule } from 'primeng/button';
+import { MenuModule } from 'primeng/menu';
+
+@Component({
+    standalone: true,
+    selector: 'app-notifications-widget',
+    imports: [ButtonModule, MenuModule],
+    templateUrl: './notificationswidget.html'
+})
+export class NotificationsWidget {
+
+}
