+++ src/app/pages/dashboard/components/progreso-por-modulos/moduleprogresswidget.html
@@ -1,0 +1,117 @@
+<div class="card">
+    <div class="flex justify-between items-center mb-6">
+        <div class="font-semibold text-xl">Progreso por módulo</div>
+    </div>
+    <ul class="list-none p-0 m-0">
+        <li class="flex flex-col md:flex-row md:items-center md:justify-between mb-6">
+            <div>
+                <span class="text-surface-900 dark:text-surface-0 font-medium mr-2 mb-1 md:mb-0">Caracteriza tu
+                    asignatura</span>
+                <div class="mt-1 text-muted-color">Momento 1</div>
+            </div>
+            <div class="mt-2 md:mt-0 flex items-center">
+                <div class="bg-surface-300 dark:bg-surface-500 rounded-border overflow-hidden w-40 lg:w-24"
+                    style="height: 8px">
+                    <div class="bg-orange-500 h-full" style="width: 80%"></div>
+                </div>
+                <span class="text-orange-500 ml-4 font-medium">%80</span>
+            </div>
+        </li>
+        <li class="flex flex-col md:flex-row md:items-center md:justify-between mb-6">
+            <div>
+                <span class="text-surface-900 dark:text-surface-0 font-medium mr-2 mb-1 md:mb-0"> Factores
+                    situacionales</span>
+                <div class="mt-1 text-muted-color">Momento 1</div>
+            </div>
+            <div class="mt-2 md:mt-0 ml-0 md:ml-20 flex items-center">
+                <div class="bg-surface-300 dark:bg-surface-500 rounded-border overflow-hidden w-40 lg:w-24"
+                    style="height: 8px">
+                    <div class="bg-cyan-500 h-full" style="width: 16%"></div>
+                </div>
+                <span class="text-cyan-500 ml-4 font-medium">%16</span>
+            </div>
+        </li>
+        <li class="flex flex-col md:flex-row md:items-center md:justify-between mb-6">
+            <div>
+                <span class="text-surface-900 dark:text-surface-0 font-medium mr-2 mb-1 md:mb-0">Ambientes Sanos y
+                    seguros</span>
+                <div class="mt-1 text-muted-color">Momento 2</div>
+            </div>
+            <div class="mt-2 md:mt-0 ml-0 md:ml-20 flex items-center">
+                <div class="bg-surface-300 dark:bg-surface-500 rounded-border overflow-hidden w-40 lg:w-24"
+                    style="height: 8px">
+                    <div class="bg-pink-500 h-full" style="width: 7%"></div>
+                </div>
+                <span class="text-pink-500 ml-4 font-medium">%7</span>
+            </div>
+        </li>
+        <li class="flex flex-col md:flex-row md:items-center md:justify-between mb-6">
+            <div>
+                <span class="text-surface-900 dark:text-surface-0 font-medium mr-2 mb-1 md:mb-0">RAP y RAC</span>
+                <div class="mt-1 text-muted-color">Momento 2</div>
+            </div>
+            <div class="mt-2 md:mt-0 ml-0 md:ml-20 flex items-center">
+                <div class="bg-surface-300 dark:bg-surface-500 rounded-border overflow-hidden w-40 lg:w-24"
+                    style="height: 8px">
+                    <div class="bg-green-500 h-full" style="width: 3%"></div>
+                </div>
+                <span class="text-primary ml-4 font-medium">%3</span>
+            </div>
+        </li>
+        <li class="flex flex-col md:flex-row md:items-center md:justify-between mb-6">
+            <div>
+                <span class="text-surface-900 dark:text-surface-0 font-medium mr-2 mb-1 md:mb-0">Actividades de
+                    aprendizaje</span>
+                <div class="mt-1 text-muted-color">Momento 3</div>
+            </div>
+            <div class="mt-2 md:mt-0 ml-0 md:ml-20 flex items-center">
+                <div class="bg-surface-300 dark:bg-surface-500 rounded-border overflow-hidden w-40 lg:w-24"
+                    style="height: 8px">
+                    <div class="bg-purple-500 h-full" style="width: 3%"></div>
+                </div>
+                <span class="text-purple-500 ml-4 font-medium">%3</span>
+            </div>
+        </li>
+        <li class="flex flex-col md:flex-row md:items-center md:justify-between mb-6">
+            <div>
+                <span class="text-surface-900 dark:text-surface-0 font-medium mr-2 mb-1 md:mb-0">Como evaluaré</span>
+                <div class="mt-1 text-muted-color">Momento 3</div>
+            </div>
+            <div class="mt-2 md:mt-0 ml-0 md:ml-20 flex items-center">
+                <div class="bg-surface-300 dark:bg-surface-500 rounded-border overflow-hidden w-40 lg:w-24"
+                    style="height: 8px">
+                    <div class="bg-teal-500 h-full" style="width: 0%"></div>
+                </div>
+                <span class="text-teal-500 ml-4 font-medium">%0</span>
+            </div>
+        </li>
+        <li class="flex flex-col md:flex-row md:items-center md:justify-between mb-6">
+            <div>
+                <span class="text-surface-900 dark:text-surface-0 font-medium mr-2 mb-1 md:mb-0">Secuencia del
+                    Curso</span>
+                <div class="mt-1 text-muted-color">Momento 4</div>
+            </div>
+            <div class="mt-2 md:mt-0 ml-0 md:ml-20 flex items-center">
+                <div class="bg-surface-300 dark:bg-surface-500 rounded-border overflow-hidden w-40 lg:w-24"
+                    style="height: 8px">
+                    <div class="bg-teal-500 h-full" style="width: 0%"></div>
+                </div>
+                <span class="text-teal-500 ml-4 font-medium">%0</span>
+            </div>
+        </li>
+        <li class="flex flex-col md:flex-row md:items-center md:justify-between mb-6">
+            <div>
+                <span class="text-surface-900 dark:text-surface-0 font-medium mr-2 mb-1 md:mb-0">Bibliografia y medios
+                    educativos</span>
+                <div class="mt-1 text-muted-color">Momento 4</div>
+            </div>
+            <div class="mt-2 md:mt-0 ml-0 md:ml-20 flex items-center">
+                <div class="bg-surface-300 dark:bg-surface-500 rounded-border overflow-hidden w-40 lg:w-24"
+                    style="height: 8px">
+                    <div class="bg-teal-500 h-full" style="width: 0%"></div>
+                </div>
+                <span class="text-teal-500 ml-4 font-medium">%0</span>
+            </div>
+        </li>
+    </ul>
+</div>
