+++ src/app/pages/dashboard/components/progreso-por-modulos/moduleprogresswidget.ts
@@ -1,0 +1,14 @@
+import { Component } from '@angular/core';
+import { CommonModule } from '@angular/common';
+
+
+@Component({
+    standalone: true,
+    selector: 'app-moduleprogress-widget',
+    imports: [CommonModule],
+    templateUrl: './moduleprogresswidget.html'
+
+})
+export class ModuleProgressWidget {
+
+}
