+++ src/app/pages/dashboard/dashboard.html
@@ -5,14 +5,10 @@
         <div class="col-span-12">
         <app-generalprogress-widget/>
     </div>
-    <!--<app-stats-widget class="contents" />-->
     <div class="col-span-12 xl:col-span-6">
-
-        <!--<app-recent-sales-widget />-->
-        <app-best-selling-widget />
+        <app-moduleprogress-widget/>
     </div>
     <div class="col-span-12 xl:col-span-6">
-        <!--<app-revenue-stream-widget />-->
         <app-notifications-widget />
     </div>
 </div>
