+++ src/app/pages/dashboard/dashboard.ts
@@ -1,16 +1,13 @@
 import { Component } from '@angular/core';
-import { NotificationsWidget } from './components/notificationswidget';
-import { StatsWidget } from './components/statswidget';
-import { RecentSalesWidget } from './components/recentsaleswidget';
-import { BestSellingWidget } from './components/bestsellingwidget';
+import { NotificationsWidget } from './components/notificaciones/notificationswidget';
 import { TimelineWidget} from './components/linea-tiempo/timelinewidget';
-import { RevenueStreamWidget } from './components/revenuestreamwidget';
 import { Generalprogresswidget } from './components/barra-progreso-general/generalprogresswidget';
+import {ModuleProgressWidget} from './components/progreso-por-modulos/moduleprogresswidget'
 
 
 @Component({
     selector: 'app-dashboard',
-    imports: [StatsWidget, RecentSalesWidget, BestSellingWidget, RevenueStreamWidget, NotificationsWidget, TimelineWidget, Generalprogresswidget],
+    imports: [NotificationsWidget, TimelineWidget, Generalprogresswidget,ModuleProgressWidget],
     templateUrl: 'dashboard.html'
 
 })
